Combine PluginRuntime, PluginContext, and BaseContext all into PluginRuntime.
