--- vm/base-context.h
@@ -1,60 +1,0 @@
-// vim: set sts=2 ts=8 sw=2 tw=99 et:
-//
-// Copyright (C) 2006-2015 AlliedModders LLC
-//
-// This file is part of SourcePawn. SourcePawn is free software: you can
-// redistribute it and/or modify it under the terms of the GNU General Public
-// License as published by the Free Software Foundation, either version 3 of
-// the License, or (at your option) any later version.
-//
-// You should have received a copy of the GNU General Public License along with
-// SourcePawn. If not, see http://www.gnu.org/licenses/.
-//
-#ifndef _INCLUDE_SOURCEPAWN_BASECONTEXT_H_
-#define _INCLUDE_SOURCEPAWN_BASECONTEXT_H_
-
-#include <sp_vm_api.h>
-
-using namespace SourcePawn;
-namespace sp {
-
-// Since the legacy API has a lot of cruft in it, we use a base class to handle common
-// code and stub deprecated functions.
-class BasePluginContext : public SourcePawn::IPluginContext
-{
-  public:
-    BasePluginContext();
-    virtual ~BasePluginContext();
-
-    // Generic API access.
-    void SetKey(int k, void* value) override;
-    bool GetKey(int k, void** value) override;
-    SourcePawn::ISourcePawnEngine2* APIv2() override;
-    void ReportError(const char* fmt, ...) override;
-    void ReportErrorVA(const char* fmt, va_list ap) override;
-    void ReportFatalError(const char* fmt, ...) override;
-    void ReportFatalErrorVA(const char* fmt, va_list ap) override;
-    void ReportErrorNumber(int error) override;
-    cell_t ThrowNativeErrorEx(int error, const char* msg, ...) override;
-    cell_t ThrowNativeError(const char* msg, ...) override;
-    int GetLastNativeError() override;
-    cell_t BlamePluginError(SourcePawn::IPluginFunction* pf, const char* msg, ...) override;
-    IFrameIterator* CreateFrameIterator() override;
-    void DestroyFrameIterator(IFrameIterator* it) override;
-
-    int LocalToArrayPtr(cell_t base, ARRAY_PTR* out) override;
-    void* GetArrayData(ARRAY_PTR handle, uint32_t* size = nullptr) override;
-
-    bool IsDebugging() override;
-
-  protected:
-    Environment* env_;
-
-  private:
-    void* m_keys[4];
-    bool m_keys_set[4];
-};
-
-} // namespace sp
-
-#endif // _INCLUDE_SOURCEPAWN_BASECONTEXT_H_
