+++ vm/interpreter.cpp
@@ -28,7 +28,7 @@
 #include "interpreter.h"
 #include "method-info.h"
 #include "pcode-reader.h"
-#include "plugin-context.h"
+#include "plugin-runtime.h"
 #include "runtime-helpers.h"
 #include "watchdog_timer.h"
