+++ vm/x64/frames-x64.h
@@ -24,7 +24,8 @@ namespace sp {
 
 using namespace SourcePawn;
 
-class PluginContext;
+class PluginRuntime;
+typedef PluginRuntime PluginContext;
 
 // We create x64 stack frames like:
 //   [return address]
